Program that parses through prereq info to extraxt course names
Input should be the coreq and prereq sections of course detail. Output will be all the "or" courses separated by "/" and "and" courses separated by "|".

Paste the output into a cell, go to Data tab, then "Split Text to Columns" and separate by "|", and apply to split "and" courses into different cells